Don't allow deleting secrets with consumers
Based on historical discussion the Barbican project
rejected changing the secret consumers implementation
to disallow deletion of secrets that has consumers
due to it being a behaviour change.

This takes a new approach and implements the API
behaviour change in a new 1.2 microversion.

As operators having end users delete secrets that
the service depend upon is a major painpoint that
we want to address. The implementation of secret
consumers should have considered enforcing this
behaviour in the API in the first place instead
of relying on clients to implement the safety
measure putting the burden on the ecosystem to
determine if a secret can be safely removed or
not we think is a bad practice.

This changes the delete secret API endpoint in
a new microversion 1.2 where it doesn't allow
deleting a secret that has consumers registered
unless the `force` query parameter with a truthy
value is given.

Change-Id: I89cd9608086084cb51a51de2b980883f1f35f189
Signed-off-by: Tobias Urdin <tobias.urdin@binero.com>